refactor: change the rating slider to use whole numbers
A warning in the logs pops up if the step size is a float, so change it to use 1 and instead use the
formatter to display the correct value. This also ensures that we have the correct value used for
the rating throughout, we just need to convert between int/float for display purposes.

+++ features/series/src/main/java/com/chesire/nekome/app/series/detail/SeriesDetailSheetFragment.kt
@@ -146,9 +146,9 @@ class SeriesDetailSheetFragment : BottomSheetDialogFragment() {
 
     private fun setupRating(model: MutableSeriesModel) {
         binding.detailRating.ratingSlider.apply {
-            value = model.rating
+            value = model.rating.toFloat()
             addOnChangeListener { _, value, _ ->
-                model.rating = value
+                model.rating = value.toInt()
             }
         }
     }
